feat: プロジェクト構造のリファクタリングとバグ修正
プロジェクト全体の構造をFlaskの標準的なディレクトリ構成にリファクタリングし、保守性と拡張性を向上させた。また、アプリケーションが起動しないバグと、APIアクセス時のコンテキストエラーを修正した。
