+++ .github/copilot-instructions.md
@@ -1,0 +1,23 @@
+# 下記を遵守してコーディングし、言語設定は日本語でレビューなどは記述してください。
+
+# 龍谷バスナビプロジェクトのためのGitHub Copilotへの指示
+
+## 1. 全般的なルール
+- コメント、docstring、およびユーザー向けのテキストは全て**日本語**で記述すること。
+- PythonコードはPEP 8スタイルガイドに従うこと。
+- 明確で簡潔なコードを記述すること。複雑なロジックにはコメントを追加して説明すること。
+
+## 2. 技術スタックとフレームワーク
+- これは**Flask**のWebアプリケーションです。
+- アプリケーションの構造化には**Flask Blueprint**を使用すること。ルート定義は`buskita/routes.py`に配置すること。
+- ビジネスロジックは`buskita/services/`ディレクトリ内のサービス関数に分離すること。
+
+## 3. API設計
+- APIエンドポイントはJSONオブジェクトを返すこと。
+- 成功時のレスポンスには`status: 'success'`キーと`data`キーを含めること。
+- エラー時のレスポンスには`status: 'error'`キーと`message`キーを含めること。
+- 成功レスポンスの例: `{"status": "success", "data": {"bus_id": 1, "latitude": 34.9, "longitude": 135.9}}`
+
+## 4. 禁止するパターン
+- 秘密鍵、APIキー、パスワードなどをソースコードに直接コミットしないこと。環境変数と`config.py`ファイルを使用すること。
+- 巨大な単一関数（モノリシック関数）の作成は避けること。複雑なタスクは再利用可能な小さな関数に分割すること。
