+++ .github/pull_request_template.md
@@ -1,0 +1,77 @@
+---
+name: 'プルリクエスト'
+about: '変更内容を記述するためのテンプレート'
+title: '[種別] 変更内容の要約 (例: [BUG] APIエラーを修正)'
+labels: ''
+assignees: ''
+
+---
+
+## 1. 関連するIssue
+
+- Closes # (ここにIssue番号を記述)
+
+---
+
+## 2. 変更の概要
+
+<!--
+このプルリクエストが解決する問題や、追加する機能の概要を簡潔に記述してください。
+-->
+
+---
+
+## 3. 主な変更点
+
+<!--
+具体的な変更内容をリスト形式で記述してください。
+例:
+- `buskita/services/__init__.py`: アプリケーションコンテキスト外で発生していたエラーを修正
+- `buskita/web_map_app.py`: ファイルを分割し、`routes.py`と`services`にリファクタリング
+- `.gitignore`: 不要になったエントリを削除
+-->
+
+- 
+- 
+- 
+
+---
+
+## 4. なぜこの変更が必要か？
+
+<!--
+変更の背景や理由を説明してください。
+バグ修正の場合は「どのような問題があったか」、機能追加の場合は「なぜこの機能が必要か」を記述します。
+-->
+
+---
+
+## 5. 動作確認の方法
+
+<!--
+レビュアーが変更内容を検証するための手順を具体的に記述してください。
+例:
+1. `pip install -r requirements.txt` を実行する
+2. `python3 run.py` を実行する
+3. ブラウザで `http://127.0.0.1:5001` を開き、地図上にバスが表示されることを確認する
+-->
+
+1. 
+2. 
+3. 
+
+---
+
+## 6. スクリーンショット（UIの変更がある場合）
+
+| Before | After |
+| :---: | :---: |
+| (変更前のスクリーンショット) | (変更後のスクリーンショット) |
+
+---
+
+## 7. レビュアーへの連絡事項
+
+<!--
+特に注意して確認してほしい点や、懸念事項などがあれば記述してください。
+-->
