--- bus_location_2025-06-27_18-19-38.json
@@ -1,5 +1,0 @@
-{
-  "isCached": true,
-  "bus": [],
-  "language": 1
-}
