+++ config.py
@@ -1,0 +1,8 @@
+API_BASE_URL = "https://api.buskita.com"
+SITE_ID = 9
+BACKUP_FILE = 'buskita/archive/last_known_buses.json' # パスをプロジェクトルートからの相対パスに変更
+HEADERS = {
+    'Accept': 'application/json',
+    'Content-Type': 'application/json',
+    'User-Agent': 'Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/605.1.15'
+}
