--- docs/REQUIREMENTS copy.md
@@ -1,41 +1,0 @@
-# 要件定義書: Ryukoku Bus Navi
-
-## 1. 概要
-
-| 項目 | 内容 |
-| :--- | :--- |
-| **アプリ名** | Ryukoku Bus Navi (仮称) |
-| **目的** | 龍谷大学瀬田キャンパスへのバス通学における、不安や不便を解消し、より快適でスマートな通学体験を提供する。 |
-| **ターゲット** | 龍谷大学 瀬田キャンパスに通う学生、及び教職員。 |
-| **プラットフォーム**| Webブラウザ（スマートフォンでの閲覧を主眼に置いたレスポンシブデザイン） |
-
-## 2. 機能要件
-
-### 2.1. 必須機能（MVP: Minimum Viable Product）
-
-まず最初に、アプリの核となる、最小限の価値を提供するための機能を定義します。
-
-| ID | 機能名 | 機能詳細 | 備考（技術仕様） |
-| :--- | :--- | :--- | :--- |
-| F-01 | **リアルタイムバスマップ** | ・瀬田キャンパス周辺の地図を表示する。<br>・帝産バスの位置を、バスのアイコンで地図上に表示する。<br>・バスの位置は3秒間隔で自動更新される。 | ・バックエンド: Flask<br>・フロントエンド: Leaflet.js<br>・API: `get-buses` (siteId: 9) |
-| F-02 | **バス情報の表示** | ・地図上のバスアイコンをタップすると、ポップアップで情報を表示する。<br>・表示項目: `路線名`, `バスID` | ・`get-buses` APIから取得できる情報を基に表示する。 |
-| F-03 | **主要ランドマークの表示**| ・地図上で「龍谷大学」と「瀬田駅」の場所を、特別なアイコンで確認できる。<br>・アイコンをタップすると、名称が表示される。 | ・API: `get-landmarks-dictionary` を利用、または座標を直接指定する。 |
-
-### 2.2. 発展機能（将来的な拡張）
-
-必須機能が完成した後、アプリをより便利で魅力的にするために追加を検討する機能です。
-
-| ID | 機能名 | 機能詳細 | 備考（技術仕様） |
-| :--- | :--- | :--- | :--- |
-| F-04 | **遅延情報の表示** | ・バスのポップアップ表示項目に「遅延情報」を追加する。<br>・例：「ほぼ定刻」「5分遅れ」 | ・API: `get-bus` を追加で呼び出し、`delayMinutes` を取得する必要がある。 |
-| F-05 | **乗車率の可視化** | ・バスの混雑度に応じて、アイコンの色を変化させる。<br>・例：緑(空席あり), 青(座れる), オレンジ(やや混雑), 赤(満員) | ・API: `get-bus` を追加で呼び出し、`occupancyStatus` を取得する必要がある。 |
-| F-06 | **時刻表機能** | ・「瀬田駅 ⇔ 龍谷大学」などの主要路線の平日・土日祝の時刻表を、見やすい形式で表示する。 | ・`ryukoku_bus_app.py` 内のデータを流用するか、最新の情報を別途取得する。 |
-| F-07 | **接近通知機能** | ・指定したバス停にバスが近づいたことを検知し、Webプッシュ通知を送信する。 | ・バックグラウンドでの位置情報監視と、Web Push APIの知識が必要。 |
-
-## 3. 非機能要件
-
-| ID | 要件名 | 詳細 |
-| :--- | :--- | :--- |
-| NF-01 | **パフォーマンス** | ・ページの初期表示は3秒以内を目指す。<br>・地図の操作（ズーム、移動）がスムーズであること。 |
-| NF-02 | **UI/UX** | ・スマートフォンでの閲覧・操作に最適化された、シンプルで直感的なデザインであること。 |
-| NF-03 | **安定性** | ・APIが一時的に不調な場合でも、アプリが完全に停止することなく、最後に取得した情報を表示するなど、堅牢な動作をすること。（バックアップ/フォールバック機能） |
