+++ run.py
@@ -1,0 +1,6 @@
+from buskita import create_app
+
+app = create_app()
+
+if __name__ == '__main__':
+    app.run(debug=True, host='0.0.0.0', port=5001)
